jarvis/requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
wikipedia==1.4.0
requests==2.31.0
psutil==5.9.5
pyjokes==0.6.0
openai==0.28.0
numpy==1.26.0
nltk==3.8.1
python-dotenv==1.0.0 